+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 01. :gift: Added interface `dev.orne.test.rnd.Generator`
 01. :gift: Added class `dev.orne.test.rnd.AbstractGenerator`
+01. :gift: Added interface `dev.orne.test.rnd.TypedGenerator`
+01. :gift: Added class `dev.orne.test.rnd.AbstractTypedGenerator`